multiple changes

+++ library/qc.py
@@ -3,24 +3,27 @@
 # This module provides QC-related functions for all Illumina sequencer types.
 
 import subprocess
-import re, os
+import re
+import os
 import shutil
+import gzip
+import operator
 from multiprocessing import Pool
 import nsc, utilities
 
 template_dir = os.path.dirname(os.path.dirname(__file__)) + "/template"
 
-def run_fastqc(files, output_dir=None, max_threads=None):
+def run_fastqc(files, demultiplex_dir, output_dir=None, max_threads=None):
     '''Run fastqc on a set of fastq files'''
     args = []
     if max_threads:
-        args += ['--threads=' + str(max_threads)]
+        args += ['--threads=' + str(max_threads), '--quiet']
     if output_dir:
         args += ["--outdir=" + output_dir]
     args += files
     print "Running fastqc on", len(files), "files"
     DEVNULL = open(os.devnull, 'wb') # discard output
-    rc = subprocess.call([nsc.FASTQC] + args, stdout=DEVNULL)
+    rc = subprocess.call([nsc.FASTQC] + args, stdout=DEVNULL, cwd=demultiplex_dir)
     
 
 
@@ -45,19 +48,38 @@ def move_fastqc_results(quality_control_dir, sample):
         pass
     
     for f in sample.files:
-        fp = f.path
-        fastqc_result_dir_name = fastqc_dir(fp)
+        if not f.empty:
+            fp = f.path
+            fastqc_result_dir_name = fastqc_dir(fp)
+    
+            
+            fastqc_output = os.path.join(quality_control_dir, fastqc_result_dir_name)
+            # Don't need the zip, as we have the uncompressed version.
+            os.remove(os.path.join(quality_control_dir, fastqc_result_dir_name + ".zip"))
+    
+            # Move from root fastqc to sample directory
+            dst_file = os.path.join(sample_dir, fastqc_result_dir_name)
+            if os.path.exists(dst_file):
+                shutil.rmtree(dst_file)
+            os.rename(fastqc_output, dst_file)
+            
 
-        # Don't need the zip, as we have the uncompressed version.
-        os.remove(os.path.join(quality_control_dir, fastqc_result_dir_name + ".zip"))
+def update_stats_fastqc(quality_control_dir, sample):
+    '''Gets total number of sequences from fastqc output, or confirms existing
+    number if already set'''
 
-        # Move from root fastqc to sample directory
-        src_file = os.path.join(quality_control_dir, fastqc_result_dir_name)
-        dst_file = os.path.join(sample_dir, fastqc_result_dir_name)
-        if os.path.exists(dst_file):
-            shutil.rmtree(dst_file)
-        os.rename(src_file, dst_file)
-
+    sample_dir = os.path.join(quality_control_dir, "Sample_" + sample.name)
+    for f in sample.files:
+        fastqc_result_dir = os.path.join(sample_dir, fastqc_dir(f.path))
+        with open(fastqc_result_dir + "/fastqc_data.txt") as fqc_data:
+            for l in fqc_data:
+                total = re.match("Total Sequences (\d+)$", l)
+                if total:
+                    if f.num_pf_reads:
+                        assert f.num_pf_reads == int(total.group(1))
+                    else:
+                        f.num_pf_reads = int(total.group(1))
+                    break
 
 
 
@@ -89,17 +111,15 @@ def generate_report_for_customer(args):
     try:
         os.mkdir(pdf_dir)
     except OSError:
-        shutil.rmtree(pdf_dir)
-        os.mkdir(pdf_dir)
-
-    os.chdir(pdf_dir) # fine to do cd in parallel since we're using multiprocessing
+        pass
 
     raw_replacements = {
         '__RunName__': tex_escape(run_id),
-        '__VersionString__': tex_escape(" & ".join([software_versions['RTA'], software_versions['bcl2fastq']])),
+        '__Programs__': tex_escape(" & ".join(v[0] for v in software_versions)),
+        '__VersionString__': tex_escape(" & ".join(v[1] for v in software_versions)),
         '__SampleName__': tex_escape(sample.name),
-        '__ReadNum__': tex_escape(str(fastq.read_num)),
-        '__TotalN__': tex_escape(str(fastq.num_pf_reads)),
+        '__ReadNum__': str(fastq.read_num),
+        '__TotalN__': utilities.display_int(fastq.num_pf_reads),
         '__Folder__': '../' + fastqc_dir(fastq.path),
         '__TemplateDir__': template_dir
             }
@@ -108,19 +128,19 @@ def generate_report_for_customer(args):
     report_root_name = ".".join((run_id, str(fastq.lane.id), "Sample_" + sample.name,
             "Read" + str(fastq.read_num), "qc"))
     fname = report_root_name + ".tex"
-    of = open(fname, "w")
+    of = open(pdf_dir + "/" + fname, "w")
     of.write(replace_multiple(replacements, template))
     of.close()
     DEVNULL = open(os.devnull, 'wb') # discard output
-    subprocess.check_call([nsc.PDFLATEX, '-shell-escape', fname], stdout=DEVNULL)
+    subprocess.check_call([nsc.PDFLATEX, '-shell-escape', fname], stdout=DEVNULL, stdin=DEVNULL, cwd=pdf_dir)
 
     pdfname = report_root_name + ".pdf"
-    shutil.copyfile(pdfname, os.path.join(fastq_dir, os.path.dirname(fastq.path), pdfname))
-    os.rename(pdfname, quality_control_dir + "/" + pdfname)
+    shutil.copyfile(pdf_dir + "/" + pdfname, os.path.join(fastq_dir, os.path.dirname(fastq.path), pdfname))
+    os.rename(pdf_dir + "/" + pdfname, quality_control_dir + "/" + pdfname)
 
 
-def compute_md5(proj_dir, threads):
-    md5data = utilities.check_output([nsc.MD5DEEP, "-rl", "-j" + str(threads), "."],
+def compute_md5(proj_dir, threads, files):
+    md5data = utilities.check_output([nsc.MD5DEEP, "-rl", "-j" + str(threads)] + files,
             cwd=proj_dir)
     open(os.path.join(proj_dir, "md5sum.txt"), "w").write(md5data)
 
@@ -192,9 +212,11 @@ def generate_internal_html_report(quality_control_dir, samples):
                 out_file.write(celln.format(subDir=subdir, fileName=fq_name, index=i))
                 
                 report_path = os.path.join(quality_control_dir, subdir, fqc_dir, "fastqc_report.html")
-                overrepresented_seq_buffer += extract_format_overrepresented(report_path, fq_name, "M" + str(i))
+                if not fq.empty:
+                    overrepresented_seq_buffer += extract_format_overrepresented(report_path, fq_name, "M" + str(i))
 
                 i += 1
+
         out_file.write("</table>\n")
         out_file.write(overrepresented_seq_buffer)
         out_file.write("</div>\n</body>\n</html>\n")
@@ -204,14 +226,15 @@ def generate_internal_html_report(quality_control_dir, samples):
 def write_sample_info_table(output_path, runid, project):
     with open(output_path, 'w') as out:
         out.write('--------------------------------		\n')
-        out.write('Email for ' + project.name)
+        out.write('Email for ' + project.name + "\n")
         out.write('--------------------------------		\n\n')
         nsamples = len(project.samples)
         out.write('Sequence ready for download - sequencing run ' + runid + ' - ' + project.name + ' (' + str(nsamples) + ' samples)\n\n')
 
-        for f in (fi for s in project.samples for fi in s.files):
+        files = sorted((fi for s in project.samples for fi in s.files), key=lambda x: os.path.basename(x.path))
+        for f in files:
             out.write(os.path.basename(f.path) + "\t")
-            out.write(str(f.num_pf_reads) + "\t")
+            out.write(utilities.display_int(f.num_pf_reads) + "\t")
             out.write("fragments\n")
 
 
@@ -223,15 +246,21 @@ def write_internal_sample_table(output_path, runid, projects):
         out.write("--------------------------------\n\n")
         out.write("Summary for run " + runid + "\n\n")
         out.write("Need not copy column A\n\n")
-
-        for s in (s for proj in projects for s in proj.samples if proj.name != "Undetermined_indices"):
+        samples = sorted(
+                (s
+                for proj in projects
+                for s in proj.samples
+                if proj.name != "Undetermined_indices"),
+                key=operator.attrgetter('name')
+                )
+        for s in samples:
             for f in s.files:
                 if f.read_num == 1:
                     out.write(s.name + "\t")
-                    out.write(str(int(round(f.lane.raw_cluster_density))) + "\t")
-                    out.write(str(int(round(f.lane.pf_cluster_density))) + "\t")
-                    out.write(str(f.percent_of_pf_clusters) + "%\t")
-                    out.write(str(int(round(f.num_pf_reads))) + "\t")
+                    out.write(utilities.display_int(f.lane.raw_cluster_density) + "\t")
+                    out.write(utilities.display_int(f.lane.pf_cluster_density) + "\t")
+                    out.write("%4.2f" % (f.percent_of_pf_clusters) + "%\t")
+                    out.write(utilities.display_int(f.num_pf_reads) + "\t")
                     out.write("ok\t\tok\n")
 
 
@@ -254,22 +283,28 @@ Lane	Project	PF cluster no	PF ratio	Raw cluster density(/mm2)	PF cluster density
 '''.format(runId = runid)
         out.write(summary_email_head)
         # assumes 1 project per lane, and undetermined
-        lane_proj = dict((proj.samples[0].files[0].lane.id, proj) for proj in projects if proj.name != "Undetermined_indices")
-        lane_undetermined = dict((proj.samples[0].files[0].lane.id, proj) for proj in projects if proj.name == "Undetermined_indices")
+        # Dict: lane ID => (lane object, project object)
+        lane_proj = dict((f.lane.id, (f.lane, proj)) for proj in projects
+                for s in proj.samples for f in s.files if proj.name != "Undetermined_indices")
+        # lane ID => file object (will be the last one)
+        lane_undetermined = dict((f.lane.id, f) for proj in projects
+                for s in proj.samples for f in s.files if proj.name == "Undetermined_indices")
 
         for l in sorted(lane_proj.keys()):
-            proj = lane_proj[l]
-            undetermined = lane_undetermined[l]
-            lane = proj.samples[0].files[0].lane
+            lane, proj = lane_proj[l]
+            undetermined_file = lane_undetermined[l]
             out.write(str(l) + "\t")
             out.write(proj.name + "\t")
-            cluster_no = sum(f.num_pf_reads for s in proj.samples + undetermined.samples for f in s.files)
-            out.write('"' + utilities.display_int(cluster_no) + '"\t')
-            out.write(str(lane.pf_ratio) + "\t")
-            out.write('"' + utilities.display_int(lane.raw_cluster_density) + '"\t')
-            out.write('"' + utilities.display_int(lane.pf_cluster_density) + '"\t')
-            undetermined_ratio = undetermined.samples[-1].files[-1].percent_of_pf_clusters
-            out.write(str(undetermined_ratio) + "%\t")
+            cluster_no = sum(f.num_pf_reads
+                    for proj in projects
+                    for s in proj.samples
+                    for f in s.files
+                    if f.lane.id == l and f.read_num == 1)
+            out.write(utilities.display_int(cluster_no) + '\t')
+            out.write("%4.2f" % (lane.pf_ratio) + "\t")
+            out.write(utilities.display_int(lane.raw_cluster_density) + '\t')
+            out.write(utilities.display_int(lane.pf_cluster_density) + '\t')
+            out.write(str(undetermined_file.percent_of_pf_clusters) + "%\t")
             out.write("ok\n")
 
 
@@ -300,7 +335,7 @@ def qc_main(input_demultiplex_dir, projects, run_id, software_versions, threads 
     output_dir = os.path.join(demultiplex_dir, "inHouseDataProcessing")
     # Unaligned/inHouseDataProcessing/QualityControl
     quality_control_dir = os.path.join(output_dir, "QualityControl")
-    delivery_dir = quality_control_dir + "/Delivery"
+    delivery_dir = output_dir + "/Delivery"
     for d in [output_dir, quality_control_dir, delivery_dir]:
         try:
             os.mkdir(d) 
@@ -308,7 +343,16 @@ def qc_main(input_demultiplex_dir, projects, run_id, software_versions, threads 
             pass
 
     print "Number of projects: ", len(projects)
-    all_fastq = [f.path for pro in projects for s in pro.samples for f in s.files]
+    all_fastq = []
+    non_empty_files = []
+
+    for f in [f for pro in projects for s in pro.samples for f in s.files]:
+        all_fastq.append(f.path)
+        gzf = gzip.open(os.path.join(demultiplex_dir, f.path))
+        f.empty = len(gzf.read(1)) == 0
+        gzf.close()
+        if not f.empty:
+            non_empty_files.append(f.path)
 
     if len(set(os.path.basename(f) for f in all_fastq)) < len(all_fastq):
         raise RuntimeError("Not all fastq file names are unique! Can't deal with this, consider splitting into smaller jobs.")
@@ -316,24 +360,28 @@ def qc_main(input_demultiplex_dir, projects, run_id, software_versions, threads 
     # Run FastQC
     # First output all fastqc results into QualityControl, them move them
     # in place later
-    os.chdir(demultiplex_dir)
-    run_fastqc(all_fastq, output_dir=quality_control_dir, max_threads=threads) 
+    run_fastqc(non_empty_files, demultiplex_dir, output_dir=quality_control_dir, max_threads=threads) 
     samples = [sam for pro in projects for sam in pro.samples]
     for s in samples:
        move_fastqc_results(quality_control_dir, s)
+       # Get number of sequences. For (Mi|Next)Seq this is the only way to 
+       # get this stat, for HiSeq this acts as a cross check.
+       update_stats_fastqc(quality_control_dir, s)
 
     # Generate PDF reports in parallel
-    # Templates at: ../template/
     template = open(template_dir + "/reportTemplate_indLane_v4.tex").read()
     arg_pack = [demultiplex_dir, quality_control_dir, run_id, software_versions, template]
     pool = Pool(int(threads))
     # Run one task for each fastq file, giving a sample reference and FastqFile as argument 
     # as well as the ones given above. Debug note: change pool.map to map for better errors.
-    pool.map(generate_report_for_customer, [tuple(arg_pack + [s,f]) for s in samples for f in s.files]) 
+    pool.map(generate_report_for_customer, [tuple(arg_pack + [s,f]) for s in samples for f in s.files if not f.empty]) 
     
     # Generate md5sums for projects
     for p in projects:
-        compute_md5(os.path.join(demultiplex_dir, p.proj_dir), threads)
+        if p.proj_dir:
+            compute_md5(os.path.join(demultiplex_dir, p.proj_dir), threads, ["."])
+        else: # Project files are in root of demultiplexing dir
+            compute_md5(demultiplex_dir, threads, ["./" + f.path for s in p.samples for f in s.files])
 
     # Generate internal reports
     generate_internal_html_report(quality_control_dir, samples)
@@ -346,7 +394,7 @@ def qc_main(input_demultiplex_dir, projects, run_id, software_versions, threads 
 
     # The following reports don't need to be used (i.e. read, copied) for  
     # LIMS-based runs, as a dedicated job will generate these for the full
-    # run after all demultiplexing has completed.
+    # run after all demultiplexing has completed (TODO).
 
     # Internal bookkeeping
     fname = delivery_dir + "/Table_for_GA_runs_" + run_id + ".xls"
